% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{determine_ylim}
\alias{determine_ylim}
\title{Helper function to determine y-limit}
\usage{
determine_ylim(
  forecasts,
  forecast_date = NULL,
  target_type,
  horizon,
  location,
  truth,
  start_at_zero = TRUE
)
}
\arguments{
\item{forecasts}{a data.frame containing forecasts from one model in the
standard long format needs to contain forecasts from different
\code{forecast_dates} to plot forecats by "horizon"}

\item{forecast_date}{the date at which forecasts were issued; if specified,
1 though 4 wk ahead forecasts are shown. Has to be NULL if horizon is
specified}

\item{target_type}{\code{"inc death"} or \code{"cum death"}}

\item{horizon}{\code{"1 wk ahead"}, \code{"2 wk ahead"}, \code{"3 wk ahead"} or
\code{"4 wk ahead"}; if specified forecasts at this horizon are plotted for
different forecast dates. Has to be NULL if forecast_date is specified}

\item{location}{the location for which to plot forecasts}

\item{truth}{the truth data set}

\item{start_at_zero}{logical (defaults to \code{TRUE}): should the y-axis include
0.}
}
\description{
Helper function to determine y-limit
}
