% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{draw_fanplot}
\alias{draw_fanplot}
\title{Draw many prediction intervals (resulting in a fanplot)}
\usage{
draw_fanplot(
  forecasts,
  target_type,
  forecast_date,
  horizon,
  location,
  levels_coverage = c(1:9/10, 0.95, 0.98),
  cols = colorRampPalette(c("deepskyblue4", "lightgrey"))(length(levels_coverage) +
    1)[-1]
)
}
\description{
Draw many prediction intervals (resulting in a fanplot)
}
