% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{draw_prediction_band}
\alias{draw_prediction_band}
\title{Add a single prediction interval}
\usage{
draw_prediction_band(
  forecasts,
  forecast_date = NULL,
  target_type,
  horizon,
  location,
  coverage,
  col = "lightgrey"
)
}
\description{
Add a single prediction interval
}
