% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{draw_truths}
\alias{draw_truths}
\title{Add smaller points for truths}
\usage{
draw_truths(truth, location, target_type)
}
\arguments{
\item{truth}{the truth data set}

\item{location}{the location for which to plot forecasts}

\item{target_type}{\code{"inc death"} or \code{"cum death"}}
}
\description{
Add smaller points for truths
}
