% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{empty_plot}
\alias{empty_plot}
\title{Create an empty plot to which forecasts can be added}
\usage{
empty_plot(xlim, ylim, xlab, ylab)
}
\description{
Create an empty plot to which forecasts can be added
}
