% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Wrap it all up into one plotting function:}
\usage{
plot_forecast(
  forecasts,
  target_type = "cum death",
  horizon = NULL,
  forecast_date = NULL,
  location,
  truth,
  levels_coverage = c(1:9/10, 0.95, 0.98),
  start = as.Date("2020-04-01"),
  end = Sys.Date() + 28,
  ylim = NULL,
  cols_intervals = colorRampPalette(c("deepskyblue4",
    "lightgrey"))(length(levels_coverage) + 1)[-1],
  col_point = "deepskyblue4",
  xlab = "date",
  ylab = target_type,
  start_at_zero = TRUE
)
}
\arguments{
\item{forecasts}{a data.frame containing forecasts from one model in the
standard long format needs to contain forecasts from different
\code{forecast_dates} to plot forecats by "horizon"}

\item{target_type}{\code{"inc death"} or \code{"cum death"}}

\item{horizon}{\code{"1 wk ahead"}, \code{"2 wk ahead"}, \code{"3 wk ahead"} or
\code{"4 wk ahead"}; if specified forecasts at this horizon are plotted for
different forecast dates. Has to be NULL if forecast_date is specified}

\item{forecast_date}{the date at which forecasts were issued; if specified,
1 though 4 wk ahead forecasts are shown. Has to be NULL if horizon is
specified}

\item{location}{the location for which to plot forecasts}

\item{truth}{the truth data set}

\item{levels_coverage}{which intervals are to be shown? Defaults to all.
\code{c(0.5, 0.95)} is a reasonable parsimonious choice.}

\item{start, end}{beginning and end of the time period to plot}

\item{ylim}{the y limits of the plot.}

\item{cols_intervals}{a vector of colors of the same length as levels_coverage}

\item{col_point}{color of the point forecasts}

\item{xlab}{a label for the x axis, defaults to a description of \code{x}.}

\item{ylab}{a label for the y axis, defaults to a description of \code{y}.}

\item{start_at_zero}{logical (defaults to \code{TRUE}): should the y-axis include
0.}
}
\description{
Wrap it all up into one plotting function:
}
