% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Run shiny app}
\usage{
run_app(...)
}
\arguments{
\item{...}{only for future-proofing, ignored for now.}
}
\description{
Run shiny app
}
\examples{
if (interactive()) {
  run_app()
}

}
