% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{truth}
\alias{truth}
\title{Truth data set containing observed cases}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7659 rows and 6 columns.
}
\source{
Automatically generated and updated by \url{https://github.com/epiforecasts/covid19-forecast-hub-europe/blob/main/viz/prepare_truth_data.R}
}
\usage{
truth
}
\description{
This truth data set contains observed cases of Covid-19 in countries covered
by the European Covid-19 Forecast Hub.
}
\keyword{datasets}
