% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{determine_ylim}
\alias{determine_ylim}
\title{Helper function to determine y-limit}
\usage{
determine_ylim(
  forecasts,
  forecast_date = NULL,
  target_type,
  horizon,
  location,
  truth,
  start_at_zero = TRUE
)
}
\description{
Helper function to determine y-limit
}
