% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_plot.R
\name{draw_fanplot}
\alias{draw_fanplot}
\title{Draw many prediction intervals (resulting in a fanplot)}
\usage{
draw_fanplot(
  forecasts,
  target_type,
  forecast_date,
  horizon,
  location,
  levels_coverage = c(1:9/10, 0.95, 0.98),
  cols = colorRampPalette(c("deepskyblue4", "lightgrey"))(length(levels_coverage) +
    1)[-1]
)
}
\arguments{
\item{forecasts}{a data.frame containing forecasts from one model in the
standard long format needs to contain forecasts from different
\code{forecast_dates} to plot forecats by "horizon"}

\item{target_type}{\code{"inc death"} or \code{"cum death"}}

\item{forecast_date}{the date at which forecasts were issued; if specified,
1 though 4 wk ahead forecasts are shown. Has to be NULL if horizon is
specified}

\item{horizon}{\code{"1 wk ahead"}, \code{"2 wk ahead"}, \code{"3 wk ahead"} or
\code{"4 wk ahead"}; if specified forecasts at this horizon are plotted for
different forecast dates. Has to be NULL if forecast_date is specified}

\item{location}{the location for which to plot forecasts}

\item{levels_coverage}{which intervals are to be shown? Defaults to all.
\code{c(0.5, 0.95)} is a reasonable parsimonious choice.}

\item{cols}{a vector of colors of the same length as levels_coverage}
}
\description{
Draw many prediction intervals (resulting in a fanplot)
}
