% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{draw_points}
\alias{draw_points}
\title{Add points for point forecasts}
\usage{
draw_points(
  forecasts,
  target_type,
  horizon,
  forecast_date,
  location,
  col = "deepskyblue4"
)
}
\description{
Add points for point forecasts
}
