% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{subset_forecasts_for_plot}
\alias{subset_forecasts_for_plot}
\title{Get the subset of a forecast file needed for plotting}
\usage{
subset_forecasts_for_plot(
  forecasts,
  forecast_date = NULL,
  target_type,
  horizon,
  location,
  type = NULL
)
}
\description{
Get the subset of a forecast file needed for plotting
}
